huge update to user permissions management - edit JSON permissions through artisan
Instead of trying to make a fancy GUI for editing permissions, I have
focused on an artisan cmdline interface instead.

It is fairly bare-bones, but is also quite powerful. At the moment,
since only "superusers" should be able to edit global permissions, this
seems like an adequate trade-off as they will likely have full SSH access to
the server anyway.